Move StrEnumMeta out of backports
